Entity + Player Animation
Entities will now walk in their set square, and when interacted with
they will face the player accordingly.

Restructured and cleaned up, ran into a minor facing bug, but I'm not
even sure if it's a bug haha.
Got some new pictures of the player too so he now "walks".